feat(DMB): add cli_debrid support in DMB
Added cli_debrid support for DMB, but this script has not been tested with the changes. Currently, DMB still defaults to running Riven, and cli_debrid needs to be enabled manually by the user in the dmb_config.json or on the DMB Frontend. Working on adding onboarding support in DMB to allow selection of either or both.